make CONTINUE_BIT and SEGMENT_BITS i32 and write u8 (not i32)

+++ protocol/src/encoding/mod.rs
@@ -1,5 +1,6 @@
-const SEGMENT_BITS: u8 = 0x7F;
-const CONTINUE_BIT: u8 = 0x80;
+// u8 perhaps?
+const SEGMENT_BITS: i32 = 0x7F;
+const CONTINUE_BIT: i32 = 0x80;
 
 pub mod varint;
 pub mod varlong;
